+++ src/main/webapp/components/result-forms/index.js
